Normalization PoC
This doesn't actually work well.
It's too easy for one or two early silence packets to get over
10% VAD P, and then normalize several early packets of silence
to a loud level.
Then, when speech actually starts, it's relatively the same
volume as the loud silence, and the VAD isn't much more confident
in the actual speech (tops out at ~30% versus as high as 70% we
see in normal operation)

Committing this for posterity, but will revert immediately.