Fix client web builds and native Magika
This was somewhat disturbing: the web example built and ran fine, but, another app on the same machine couldn't run on web due to the fonnx.dart onnx/ort.dart export.

Similarly, the native version was working until a flutter clean, at which point it became clear the **entire** file bytes were sent for inference, instead of the feature bytes (i.e. a subset of 1536 bytes)