Ensure reversed text matches original
Previously, we optimized by calling removeDiacritics + toLowerCase on the entire input string.

However, this leads to the text returned for an embedding being the lowercase + diacritics removed version.

This is unacceptable for use cases where the text is used for RAG, i.e. we want to send the original text to the LLM, not the lowercase version. It doesn't seem to affect the LLM, however, it looks weird in app UI showing what was sent.